Refactor: Move function pydoctor.test.partialclass to pydoctor.utils.partialclass
